Modifiqué el a double  las propiedades que estaban en decimal

+++ Shared/Models/Articulos.cs
@@ -13,14 +13,14 @@ namespace Shared.Models
 
 		[Range(0.01, 1000000, ErrorMessage = "Ingrese un valor mayor a 0.01 y menor o igual a 1000000")]
 		[Required(ErrorMessage = "Costo obligatorio")]
-		public decimal Costo { get; set; }
+		public double Costo { get; set; }
 
 		[Range(0, 100, ErrorMessage = "Solo se permite de 0 a 100")]
 		[Required(ErrorMessage = "Ganancia obligatoria")]
-		public decimal Ganancia { get; set; }
+		public double Ganancia { get; set; }
 
 		[Range(0.01, 1000000, ErrorMessage = "Ingrese un valor mayor a 0.01 y menor o igual a 1000000")]
 		[Required(ErrorMessage = "Precio obligatorio")]
-		public decimal Precio { get; set; }
+		public double Precio { get; set; }
 	}
 }
